+++ debug.log
@@ -619,3 +619,72 @@
 [0120/174010.731:ERROR:crash_report_database_win.cc(469)] failed to stat report
 [0120/174010.731:ERROR:crash_report_database_win.cc(469)] failed to stat report
 [0120/174010.731:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.877:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.877:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.877:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.877:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.877:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.877:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.878:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.878:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.878:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.878:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.878:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.878:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.878:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.878:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.878:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.878:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.878:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.878:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.878:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.878:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.878:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.878:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.878:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.878:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.879:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.879:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.879:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.879:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.879:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.879:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.879:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.879:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.879:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.879:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.879:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.879:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.879:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.879:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.879:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.879:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.879:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.879:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.880:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.880:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.880:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.880:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.880:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.880:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.880:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.880:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.881:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.881:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.881:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.881:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.881:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.881:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.881:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.881:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.881:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.881:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.881:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.881:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.881:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.881:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.881:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.881:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.881:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.881:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0120/205815.881:ERROR:crash_report_database_win.cc(469)] failed to stat report
