+++ src/components/create-visualization-component/CreateVisualizationContainer.tsx
@@ -1,6 +1,6 @@
 import { IState } from "../../reducers"
 import { connect } from "react-redux"
-import { SearchCurriculumComponent } from "./SearchCurriculumComponent"
+import { SearchCurriculumComponent } from "./CreateVisualizationComponent"
 import { postSubmitVisualization } from "../../action-mappers/visualization-action-mapper"
 
 const mapStateToProps = (state: IState) => {
